Move logout tests to new file
