--- tests/integration/web/auth/logout_test.py
@@ -1,53 +1,0 @@
-from nav.web.auth import logout
-from nav.web.auth.utils import ACCOUNT_ID_VAR, set_account
-from nav.web.auth.sudo import sudo
-
-
-def test_non_sudo_logout_should_remove_session_data(db, session_request, admin_account):
-    # login with admin acount
-    set_account(session_request, admin_account)
-    logout(session_request)
-    assert not hasattr(session_request, 'account')
-    assert ACCOUNT_ID_VAR not in session_request.session
-
-
-def test_non_sudo_logout_should_return_path_to_index(
-    db, session_request, admin_account
-):
-    # login with admin acount
-    set_account(session_request, admin_account)
-    result = logout(session_request)
-    assert result == '/'
-
-
-def test_sudo_logout_should_set_session_to_original_user(
-    db, session_request, admin_account, account
-):
-    # login with admin acount
-    set_account(session_request, admin_account)
-    sudo(session_request, account)
-    assert session_request.account is account
-    result = logout(session_request, sudo=True)
-    assert result == '/'
-    assert session_request.account == admin_account
-
-
-def test_non_sudo_logout_should_change_session_id(db, session_request, admin_account):
-    # login with admin acount
-    set_account(session_request, admin_account)
-    pre_session_id = session_request.session.session_key
-    logout(session_request)
-    post_session_id = session_request.session.session_key
-    assert post_session_id != pre_session_id
-
-
-def test_sudo_logout_should_change_session_id(
-    db, session_request, admin_account, account
-):
-    # login with admin acount
-    set_account(session_request, admin_account)
-    sudo(session_request, account)
-    pre_session_id = session_request.session.session_key
-    logout(session_request, sudo=True)
-    post_session_id = session_request.session.session_key
-    assert post_session_id != pre_session_id
